+++ wp-config-development.php
@@ -16,3 +16,5 @@ if ( ! defined( 'WP_DEBUG' ) ) {
 // Configure S3 Uploads.
 define( 'S3_UPLOADS_BUCKET', 'media-pri-dev/s3fs-public' );
 define( 'S3_UPLOADS_REGION', 'us-east-1' );
+
+define ( 'GRAPHQL_DEBUG', true );
