+++ wp-content/mu-plugins/the-world-site-config/configs/global/global-plugins.php
@@ -23,6 +23,7 @@ define(
 		'tw-disable-yoast-indexables/tw-disable-yoast-indexables.php',
 		'tw-endpoint-helper/tw-endpoint-helper.php',
 		'tw-episodes/tw-episodes.php',
+		'tw-graphql/tw-graphql.php',
 		'tw-import-post-types/tw-import-post-types.php', // This is the plugin that creates the custom post types for the import. Can be removed after the import is complete.
 		'tw-media/tw-media.php',
 		'tw-menus/tw-menus.php',
